requirements.txt
appdirs==1.4.4
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
distlib==0.3.1
Django==2.2
django-bootstrap4==2.2.0
django-filter==2.3.0
-e git+https://github.com/pawel-123/member_portal.git@98283dc13e5899e7baa18897b2c7f75287b5a4e9#egg=django_multiupload&subdirectory=django-multiupload-master
django-tables2==2.3.1
filelock==3.0.12
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
pandas==1.0.3
pipenv==2020.6.2
plotly==4.6.0
py-dateutil==2.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.25
virtualenv-clone==0.5.4
